+++ src/App.js
@@ -2,9 +2,7 @@ import React, { Component } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import './App.css';
 import Homepage from "./components/Homepage/Homepage";
-import InterviewPage from "./components/InterviewPage/InterviewPage";
 import LoginPage from "./components/LoginPage/LoginPage";
-import UploadPage from "./components/UploadPage/Upload";
 
 class App extends Component {
   render() {
