--- src/components/InterviewCard/interviewCard.css
@@ -1,3 +1,0 @@
-.interviewCard {
-    background-color: grey;
-}
