--- src/components/InterviewPage/mylogic.js
@@ -1,43 +1,0 @@
-var easyrtc = require('easyrtc');
-//import easyrtc from 'easyrtc';
-
-exports.my_init = function() {
-         easyrtc.setRoomOccupantListener( loggedInListener);
-         easyrtc.easyApp("Company_Chat_Line", "self", ["caller"],
-             function(myId) {
-                console.log("My easyrtcid is " + myId);
-             }
-         );
-     }
-
-
-function loggedInListener(roomName, otherPeers) {
-    var otherClientDiv = document.getElementById('otherClients');
-    while (otherClientDiv.hasChildNodes()) {
-        otherClientDiv.removeChild(otherClientDiv.lastChild);
-    }
-    for(var i in otherPeers) {
-        var button = document.createElement('button');
-        button.onclick = function(easyrtcid) {
-            return function() {
-                performCall(easyrtcid);
-            }
-        }(i);
-
-        let label = document.createTextNode(i);
-        button.appendChild(label);
-        otherClientDiv.appendChild(button);
-    }
-}
-
-
-function performCall(easyrtcid) {
-    easyrtc.call(
-       easyrtcid,
-       function(easyrtcid) { console.log("completed call to " + easyrtcid);},
-       function(errorMessage) { console.log("err:" + errorMessage);},
-       function(accepted, bywho) {
-          console.log((accepted?"accepted":"rejected")+ " by " + bywho);
-       }
-   );
-}
