--- src/firebase-debug.log
@@ -1,68 +1,0 @@
-[debug] [2019-02-12T23:26:48.363Z] ----------------------------------------------------------------------
-[debug] [2019-02-12T23:26:48.367Z] Command:       /home/arkuhn/.nvm/versions/node/v6.14.4/bin/node /home/arkuhn/.nvm/versions/node/v6.14.4/bin/firebase deploy
-[debug] [2019-02-12T23:26:48.368Z] CLI Version:   6.3.1
-[debug] [2019-02-12T23:26:48.368Z] Platform:      linux
-[debug] [2019-02-12T23:26:48.368Z] Node Version:  v6.14.4
-[debug] [2019-02-12T23:26:48.369Z] Time:          Tue Feb 12 2019 18:26:48 GMT-0500 (STD)
-[debug] [2019-02-12T23:26:48.369Z] ----------------------------------------------------------------------
-[debug] 
-[debug] [2019-02-12T23:26:48.382Z] > command requires scopes: ["email","openid","https://www.googleapis.com/auth/cloudplatformprojects.readonly","https://www.googleapis.com/auth/firebase","https://www.googleapis.com/auth/cloud-platform"]
-[debug] [2019-02-12T23:26:48.383Z] > authorizing via signed-in user
-[debug] [2019-02-12T23:26:48.383Z] [iam] checking project presencevr-ak for permissions ["firebase.projects.get","firebasehosting.sites.update"]
-[debug] [2019-02-12T23:26:48.386Z] >>> HTTP REQUEST POST https://cloudresourcemanager.googleapis.com/v1/projects/presencevr-ak:testIamPermissions  
-
-[debug] [2019-02-12T23:26:49.007Z] <<< HTTP RESPONSE 200
-[debug] [2019-02-12T23:26:49.012Z] >>> HTTP REQUEST GET https://firebase.googleapis.com/v1beta1/projects/presencevr-ak  
- 
-[debug] [2019-02-12T23:26:49.399Z] <<< HTTP RESPONSE 200
-[info] 
-[info] === Deploying to 'presencevr-ak'...
-[info] 
-[info] i  deploying hosting
-[debug] [2019-02-12T23:26:49.411Z] >>> HTTP REQUEST POST https://firebasehosting.googleapis.com/v1beta1/sites/presencevr-ak/versions  
-
-[debug] [2019-02-12T23:26:51.036Z] <<< HTTP RESPONSE 200
-[info] i  hosting[presencevr-ak]: beginning deploy...
-[info] i  hosting[presencevr-ak]: found 44 files in build
-[debug] [2019-02-12T23:26:51.082Z] [hosting] hash cache [YnVpbGQ] load error: lines is not defined
-[debug] [2019-02-12T23:26:52.426Z] [hosting] hash cache [YnVpbGQ] stored for 44 files
-[debug] [2019-02-12T23:26:52.426Z] [hosting][hash queue][FINAL]
-[debug] [2019-02-12T23:26:52.427Z] >>> HTTP REQUEST POST https://firebasehosting.googleapis.com/v1beta1/sites/presencevr-ak/versions/7cebc1646371d04d:populateFiles  
- <request body omitted>
-[debug] [2019-02-12T23:26:54.937Z] <<< HTTP RESPONSE 200
-[debug] [2019-02-12T23:26:54.949Z] [hosting][populate queue][FINAL]
-[debug] [2019-02-12T23:26:54.949Z] [hosting] uploads queued: 2
-[debug] [2019-02-12T23:27:48.758Z] ----------------------------------------------------------------------
-[debug] [2019-02-12T23:27:48.763Z] Command:       /home/arkuhn/.nvm/versions/node/v6.14.4/bin/node /home/arkuhn/.nvm/versions/node/v6.14.4/bin/firebase deploy
-[debug] [2019-02-12T23:27:48.763Z] CLI Version:   6.3.1
-[debug] [2019-02-12T23:27:48.763Z] Platform:      linux
-[debug] [2019-02-12T23:27:48.763Z] Node Version:  v6.14.4
-[debug] [2019-02-12T23:27:48.764Z] Time:          Tue Feb 12 2019 18:27:48 GMT-0500 (STD)
-[debug] [2019-02-12T23:27:48.764Z] ----------------------------------------------------------------------
-[debug] 
-[debug] [2019-02-12T23:27:48.778Z] > command requires scopes: ["email","openid","https://www.googleapis.com/auth/cloudplatformprojects.readonly","https://www.googleapis.com/auth/firebase","https://www.googleapis.com/auth/cloud-platform"]
-[debug] [2019-02-12T23:27:48.779Z] > authorizing via signed-in user
-[debug] [2019-02-12T23:27:48.779Z] [iam] checking project presencevr-ak for permissions ["firebase.projects.get","firebasehosting.sites.update"]
-[debug] [2019-02-12T23:27:48.784Z] >>> HTTP REQUEST POST https://cloudresourcemanager.googleapis.com/v1/projects/presencevr-ak:testIamPermissions  
-
-[debug] [2019-02-12T23:27:54.220Z] <<< HTTP RESPONSE 200
-[debug] [2019-02-12T23:27:54.222Z] >>> HTTP REQUEST GET https://firebase.googleapis.com/v1beta1/projects/presencevr-ak  
- 
-[debug] [2019-02-12T23:27:54.663Z] <<< HTTP RESPONSE 200
-[info] 
-[info] === Deploying to 'presencevr-ak'...
-[info] 
-[info] i  deploying hosting
-[debug] [2019-02-12T23:27:54.673Z] >>> HTTP REQUEST POST https://firebasehosting.googleapis.com/v1beta1/sites/presencevr-ak/versions  
-
-[debug] [2019-02-12T23:27:55.854Z] <<< HTTP RESPONSE 200
-[info] i  hosting[presencevr-ak]: beginning deploy...
-[info] i  hosting[presencevr-ak]: found 44 files in build
-[debug] [2019-02-12T23:27:55.911Z] [hosting] hash cache [YnVpbGQ] load error: lines is not defined
-[debug] [2019-02-12T23:27:57.281Z] [hosting] hash cache [YnVpbGQ] stored for 44 files
-[debug] [2019-02-12T23:27:57.281Z] [hosting][hash queue][FINAL]
-[debug] [2019-02-12T23:27:57.281Z] >>> HTTP REQUEST POST https://firebasehosting.googleapis.com/v1beta1/sites/presencevr-ak/versions/68c308ee366b093a:populateFiles  
- <request body omitted>
-[debug] [2019-02-12T23:27:59.031Z] <<< HTTP RESPONSE 200
-[debug] [2019-02-12T23:27:59.044Z] [hosting][populate queue][FINAL]
-[debug] [2019-02-12T23:27:59.045Z] [hosting] uploads queued: 2
